Restructured Jobs samples.
